unify: introduce `CxlTycker`
